Arreglo bugs visuales y mejoras en los messegue boxs

+++ project_final_attempt/Forms/plantilla_inicial.cs
@@ -13,7 +13,19 @@ namespace project_final_attempt
             fh.TopLevel = false;
             fh.Dock = insidePanel.Dock;
             this.insidePanel.Controls.Add(fh);
-            this.insidePanel.Tag = fh;
+
+            fh.Show();
+        }
+
+        private void Abrirform_(object form)
+        {
+            if (this.insidePanel.Controls.Count > 0)
+                this.insidePanel.Controls.Clear();
+
+            Form fh = form as Form;
+            fh.TopLevel = false;
+            insidePanel.Dock = fh.Dock;
+            this.insidePanel.Controls.Add(fh);
 
             fh.Show();
         }
@@ -29,7 +41,7 @@ namespace project_final_attempt
 
         private void btnPersonajes_Click(object sender, EventArgs e)
         {
-            Abrirform(new visualizar_personajes());
+            Abrirform_(new visualizar_personajes());
         }
 
 
@@ -48,23 +60,27 @@ namespace project_final_attempt
 
         private void btnPeliculas_Click(object sender, EventArgs e)
         {
-            Abrirform(new visualizar_peliculas());
+            Abrirform_(new visualizar_peliculas());
         }
 
 
-        private void siticoneControlBox1_Click(object sender, EventArgs e)
+        private void btnCerrar_Click(object sender, EventArgs e)
         {
-            File.Copy(@".\data\DatosPeliculas.txt", @".\data\Backup\Backup-datosPeliculas.txt", true);
-            File.Copy(@".\data\DatosPersonaje.txt", @".\data\Backup\Backup-datosPersonajes.txt", true);
-            string path_backup = @".\data\Backup\Backup-imagenes";
-            string path_imagenes = Directory.GetCurrentDirectory() + @".\data\imagenes";
-            if (!Directory.Exists(path_backup))
+            if (MessageBox.Show("¿Seguro que deseas salir del programa, la copia de seguridad se generará automáticamente?", "Advertencia", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == DialogResult.OK)
             {
-                Directory.CreateDirectory(path_backup);
-            }
-            foreach (var file in Directory.GetFiles(path_imagenes))
-            {
-                File.Copy(file, Path.Combine(path_backup, Path.GetFileName(file)), true);
+                File.Copy(@".\data\DatosPeliculas.txt", @".\data\Backup\Backup-datosPeliculas.txt", true);
+                File.Copy(@".\data\DatosPersonaje.txt", @".\data\Backup\Backup-datosPersonajes.txt", true);
+                string path_backup = @".\data\Backup\Backup-imagenes";
+                string path_imagenes = Directory.GetCurrentDirectory() + @".\data\imagenes";
+                if (!Directory.Exists(path_backup))
+                {
+                    Directory.CreateDirectory(path_backup);
+                }
+                foreach (var file in Directory.GetFiles(path_imagenes))
+                {
+                    File.Copy(file, Path.Combine(path_backup, Path.GetFileName(file)), true);
+                }
+                Application.Exit();
             }
         }
     }
